feat(editor): add 12.5 grid size option; auto-disable snap when grid is None
- GridSpacing type changed int → double in AppSettings and MainViewModel
- EditorCanvas RebuildGrid/HRuler/VRuler loops use double; modulo check uses Math.Abs < 0.001
- SnapX/SnapY use explicit double division for precision
- PageEditorOverlay ComboBox gains Grid: 12.5px item (index 1); switch cases shifted +1
- Selecting None now also sets VM.SnapToGrid = false

+++ Core/AppSettings.cs
@@ -5,7 +5,7 @@ public class AppSettings
     // Editor display toggles
     public bool   ShowGrid           { get; set; } = true;
     public bool   ShowRulers         { get; set; } = true;
-    public int    GridSpacing        { get; set; } = 100;
+    public double GridSpacing        { get; set; } = 100;
     public bool   SnapToGrid         { get; set; } = false;
     public bool   ShowSafeBoundaries { get; set; } = false;
